Set Response.reason to an empty string when obs-text characters are found in the reason phrase (#84)
This avoids parsing failures when servers return accented ISO 8859-1 characters in the reason-phrase by returning an
empty string as a fallback.

A previous patch changed the signatures involved to return the reason phrase as a `&[u8]` in order
to avoid loss of information when obs-text characters are found. However, that would require a
breaking semver change, and after discussion in #84 we decided this was an acceptable compromise.